Update project configuration and enhance component structure
- Added `@types/node` to `package.json` for improved TypeScript support.
- Updated `pnpm-lock.yaml` to reflect the new dependency.
- Configured `tsconfig.app.json` and `tsconfig.node.json` with base URL and path aliases for cleaner imports.
- Refactored `vite.config.ts` and `vitest.config.ts` to include path aliases for better module resolution.
- Moved `App` component to a new directory structure for improved organization and updated import paths accordingly.
- Enhanced test coverage with new tests for calculator functionality in `calculators.test.ts`.
- Introduced new domain logic in `calculators.ts` and `tax-calculator.ts` for tax calculations.
- Improved localization support in `es.ts` for better user experience.
