+++ src/App.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import App from "./App";
+import App from "@/app/App";
 
 describe("App integration", () => {
   beforeEach(() => {
