+++ src/components/ui/tabs.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import * as TabsPrimitive from "@radix-ui/react-tabs";
-import { cn } from "../../lib/utils";
+import { cn } from "@/utils/cn";
 
 const Tabs = TabsPrimitive.Root;
