+++ src/domain/tax/calculators.ts
@@ -1,4 +1,4 @@
-import { UVT_2026 } from "./index";
+import { UVT_2026 } from "@/domain/tax/constants";
 
 export type CalculatorId =
   | "withholdingAdvanced"
