+++ src/domain/tax/constants.ts
@@ -1,0 +1,1 @@
+export const UVT_2026 = 52374;
