+++ src/main.tsx
@@ -6,8 +6,8 @@ import "@fontsource/inter/500.css";
 import "@fontsource/inter/600.css";
 import "@fontsource/inter/700.css";
 import "@fontsource/inter/800.css";
-import "./index.css";
-import App from "./App.tsx";
+import "@/styles/index.css";
+import App from "@/app/App";
 
 registerSW({ immediate: true });
